+++ src/main/java/com/youzan/nsq/client/exception/NSQPubFailedException.java
@@ -1,0 +1,11 @@
+package com.youzan.nsq.client.exception;
+
+/**
+ * Created by lin on 17/6/6.
+ */
+public class NSQPubFailedException extends NSQException {
+
+    public NSQPubFailedException(String message) {
+        super(message);
+    }
+}