+++ src/test/java/it/youzan/nsq/client/ITTagConsumer.java
@@ -9,6 +9,7 @@ import com.youzan.nsq.client.entity.NSQMessage;
 import com.youzan.nsq.client.exception.NSQException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.testng.Assert;
 import org.testng.annotations.Test;
 
 import java.util.concurrent.CountDownLatch;
@@ -23,8 +24,9 @@ public class ITTagConsumer {
 
     @Test
     public void test() throws NSQException, InterruptedException {
-        final CountDownLatch latch = new CountDownLatch(10000);
-        final AtomicInteger received = new AtomicInteger(0);
+        final CountDownLatch latch = new CountDownLatch(200);
+        final AtomicInteger receivedTag1 = new AtomicInteger(0);
+        final AtomicInteger receivedTag2 = new AtomicInteger(0);
         NSQConfig config = new NSQConfig("default");
         config.setLookupAddresses("qabb-qa-nsqtest0:4161");
         config.setConsumerDesiredTag(new DesiredTag("TAG2"));
@@ -33,12 +35,12 @@ public class ITTagConsumer {
             public void process(NSQMessage message) {
                 logger.info("Message received: " + message.getReadableContent());
                 logger.info("message tag: " + message.getTag().toString());
-//                received.incrementAndGet();
-//                latch.countDown();
+                receivedTag1.incrementAndGet();
+                latch.countDown();
             }
         });
         consumer.setAutoFinish(true);
-        consumer.subscribe("tag2Par2Rep");
+        consumer.subscribe("tag_java_2par_2rep");
         consumer.start();
 
         NSQConfig configTag = new NSQConfig("default");
@@ -49,15 +51,16 @@ public class ITTagConsumer {
             public void process(NSQMessage message) {
                 logger.info("Message received: " + message.getReadableContent());
                 logger.info("message tag: " + message.getTag().toString());
-                received.incrementAndGet();
+                receivedTag2.incrementAndGet();
                 latch.countDown();
             }
         });
         consumerTag.setAutoFinish(true);
-        consumerTag.subscribe("tag2Par2Rep");
+        consumerTag.subscribe("tag_java_2par_2rep");
         consumerTag.start();
 
-        latch.await(10, TimeUnit.MINUTES);
-        logger.info("Consumer received {} messages.", received.get());
+        Assert.assertTrue(latch.await(5, TimeUnit.MINUTES));
+        Assert.assertEquals(receivedTag1.get(), 100);
+        Assert.assertEquals(receivedTag2.get(), 100);
     }
 }